Adding DOW offset setting
This setting allows making the day of week switch at a different time
than midnight. The thinking here is that some folks get up a bit too
early and see "Oh, it's wednesday" when really it's super early
wednesday and they need more sleep.

If this setting is set to "5", then the clock will read "Tuesday" until
5am, and then it switches to "Wednesday."

Default: 0 (disabled).